pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cogent-agent"
version = "0.1.0"
description = "Docker-containerized AI Agent using Claude Agent SDK with NATS communication"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Cogent", email = "cogent@example.com" }
]
keywords = [
    "ai",
    "agent",
    "claude",
    "anthropic",
    "nats",
    "docker",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "claude-agent-sdk>=0.1.9",
    "nats-py>=2.12.0",
    "websockets>=15.0",
    "anyio>=4.11.0",
    "aiofiles>=25.1.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.0",
    "click>=8.3.0",
    "rich>=14.0.0",
    "prompt-toolkit>=3.0.50",
    "httpx>=0.28.0",
    "structlog>=25.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=7.0.0",
    "black>=25.0.0",
    "ruff>=0.14.0",
    "mypy>=1.18.0",
]

[project.scripts]
cogent = "src.cli.main:main"
cogent-agent = "src.agent.main:main"

[project.urls]
Homepage = "https://github.com/cogent/cogent-agent"
Documentation = "https://github.com/cogent/cogent-agent#readme"
Repository = "https://github.com/cogent/cogent-agent"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
